updating movie objects

+++ app/src/main/java/com/alkss/moviecatalog/core/domain/model/local/Movie.kt
@@ -10,6 +10,6 @@ data class Movie(
     val overview: String,
     val posterPath: String,
     val releaseDate: String,
-    val rating: Float,
+    val rating: Double,
     val isFavorite: Boolean = false
 )
